html/semantics/interactive-elements/the-dialog-element/dialog-close.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>dialog element: close()</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="http://www.w3.org/html/wg/drafts/html/master/#the-dialog-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<dialog id="d1">
  <p>foobar</p>
  <button>OK</button>
</dialog>
<dialog id="d2" open>
  <p>foobar</p>
  <button>OK</button>
</dialog>
<script>
  var d1 = document.getElementById('d1'),
      d2 = document.getElementById('d2'),
      t = async_test("close() throws a close event");

  d2.onclose = t.step_func_done(function(e) {
    assert_true(e.isTrusted, "close event is trusted");
    assert_false(e.bubbles, "close event doesn't bubble");
    assert_false(e.cancelable, "close event is not cancelable");
  });

  test(function(){
    assert_throws("INVALID_STATE_ERR", function() {
      d1.close();
    });
  }, "close() on a <dialog> that doesn't have an open attribute throws an InvalidStateError exception");

  test(function(){
    try {
      assert_true(d2.open);
      assert_equals(d2.returnValue, "");
      d2.close("closedialog");
      assert_false(d2.hasAttribute("open"));
      assert_equals(d2.returnValue, "closedialog");
    } catch(e) {
      assert_unreached("An exception has been thrown");
    }
  }, "close() removes the open attribute and set the returnValue to the first argument");
</script>
